Feat : SignIn to Pairing - navigation 로직 추가
